res/build.sh: use lowdown instead of pandoc for mandoc output.
Given that pandoc is a haskell program, depending on it is bad due to
bootstrap and general platform compatibility concerns. Lowdown output
looks just as good if not better.

The ideal build will have the git repository available to it; builds
without the git repo will use SOURCE_DATE_EPOCH from the environment.