xbps/mirrors: Move Clarkson from Tier 1 to Tier 2
Clarkson reliably has sync problems and has failed to meet the quality
requirement to maintain status as a Tier 1 mirror.  As they are not
directly operated or overseen by the Void team they also don't really
satisfy the documented requirements.  As such Clarkson is now
de-listed from tier 1 status.  They remain as a Tier 2 mirror until
such time as they are not reliable enough to maintain that status.