res/void-docs: don't show picker for only 1 result
Makes for weird UX. This was implemented for fzf and skim with their
'-1' flag, but it's pretty simple to add a general implementation that
works for all pickers, including pick(1), which is what we ship with the
void-docs-browse package. In addition, we can simplify the code by
using head(1) as the fallback 'picker'.

Bug reported and fixed by mason from IRC.