ftp.acc.umu.se moves to ftp.accum.se
Background: Umeå University has decided that non-official services are
no longer allowed to exist as a sub domain to umu.se. As a consequence
of this we are moving all services from the acc.umu.se domain.

Additionally, this commit removes the duplicate entries cdimage.debian.org and
ftp.sunet.se which are all served by our mirror.