config/network/index: Also change the link to supervise symlink
In the section talking about interface-specific dhcpcd, the run script is
modified with sed to have the proper interface name, but there is also
the supervise symlink, pointing into /run/runit.

The target is "supervise.dhcpcd-eth0", which kind of works as-is, until
you also have an actual eth0 interface that is handled by a dhcpcd
supervised by runit.

So to avoid having this misleading symlink, change it to a proper target:
"supervise.dhcpcd-enp3s0" in the sample code.

Signed-off-by: Vincent Legoll <vincent.legoll@gmail.com>