config/media/pipewire: rewrite for clarity
it seems a lot of people have trouble with this guide, and there's a lot
of forks/options at each step. This rewrites the guide to streamline it
into "basic setup" (wireplumber and optionally pw-pulse) and "optional
setup" (everything else).

Also, this only documents the conf.d method as it's the most
straightforward (IMO) and provides consistent results (wrt launch order
and ensuring things are launched). I don't think we need to document
every single option and overwhelm users, this guide is already very verbose.