Force update mdBook theme files
This commit is split in two: 1) overwrite the old theme files 2) reapply
void-docs' changes to the theme. The next commit applies void-docs
changes. This split is done to make the git history clean and to clearly
showcase which changes need to be done in the second commit.

mdBook v0.4.43 was used to generate the default theme.

This update has two benefits:
1. Features added in updates to mdBook made since the old theme files
   were commited will become available.

   It is possible to "backport" newer features into the old theme files,
   but this should be a cleaner solution.
2. It should fix building void-docs using newer versions of mdBook.

   The currently packaged version of mdBook (v0.4.43) has minor
   rendering issues.